feat:adiciona exceptional de delete

+++ src/main/java/com/userTest/demo/services/ClientService.java
@@ -4,10 +4,15 @@ import com.userTest.demo.dto.ClientDto;
 import com.userTest.demo.dto.FormClientDto;
 import com.userTest.demo.entities.Client;
 import com.userTest.demo.repositories.ClientRepository;
+import com.userTest.demo.services.exceptions.DatabaseException;
+import com.userTest.demo.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -18,7 +23,7 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDto getById(Long id) {
-        var client  = repository.findById(id).get();
+        var client  = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ClientDto(client);
     }
 
@@ -37,10 +42,14 @@ public class ClientService {
 
     @Transactional
     public ClientDto update(Long id, FormClientDto dto) {
-        Client client = repository.getReferenceById(id);
-        copyDtoToEntity(dto, client);
-        client = repository.save(client);
-        return new ClientDto(client);
+        try {
+            Client client = repository.getReferenceById(id);
+            copyDtoToEntity(dto, client);
+            client = repository.save(client);
+            return new ClientDto(client);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
     private void copyDtoToEntity(FormClientDto dto, Client client) {
@@ -51,8 +60,16 @@ public class ClientService {
         client.setChildren(dto.getChildren());
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if(repository.existsById(id)) {
+            throw new ResourceNotFoundException("Id não encontrado");
+        }
+
+        try {
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 }
